requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
pillow
moviepy
tqdm
streamlit-lottie
streamlit-firebase-authentication